added test for configure item
