+++ eigami/View/MovieListViewController.swift
@@ -78,17 +78,6 @@ fileprivate extension MovieListViewController {
 
 fileprivate extension MovieListViewController {
     
-    class MovieListDataProvider: RxCollectionViewSectionedReloadDataSource<Group<Movie>> {
-        convenience init() {
-            self.init(
-                configureCell: { (ds, cv, ip, item) -> UICollectionViewCell  in
-                    let cell = cv.dequeueReusableCell(withReuseIdentifier: MovieListCell.identifier, for: ip) as! MovieListCell
-                    cell.configure(forItem: item)
-                    return cell
-            })
-        }
-    }
-    
     func bindRx() {
         searchBar.rx.text.orEmpty
             .throttle(500, scheduler: scheduler)
